move call to nodemanager in compliance ingestion to publisher
